requirements.txt
asgiref==3.7.2
Django==4.2.7
django-clearcache==1.2.1
django-timesince==2021.5.27
django-widget-tweaks==1.5.0
exrex==0.11.0
fontawesomefree==6.5.1
Pillow==10.1.0
psycopg2-binary==2.9.9
requests-futures==1.0.1
sqlparse==0.4.4
stem==1.8.2
torrequest==0.1.0
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
